combine all input components into one
When the newCase  sub-components was individualize, I wasn't able to
pass data from the select components up two levels to the parent
(NewCase & App). I was able to pass data up two levels with the
inputCase component. The solution is to combine the four sub-components
and pass data up in one bundle.