User can update the status of a case
- created a callback function from the App component to open the
ChangeStatus component and save the selected case fileNumber to state
- created a callback function from the App component to update the
selected case, (based on the fileNumber saved to state, startDate with
today's date and status with the status retrieved from the ChangeStatus
component. The ChangeStatus component is then closed and CaseManagment
reopen.
- Setup the ChangeStatus component to open and close based on the App
component state.
- Added a onClick event to each case to open the ChangeStatus component.
- Updatead the ChangeStatus component to record the new status and send
the information back with a callback function.